requirements.txt
beautifulsoup4==4.12.2
blis==0.7.10
cachetools==4.2.4
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.1
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.12.2
fsspec==2023.6.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.48.2
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-pkuseg==0.0.32
srsly==2.4.7
stanfordcorenlp==3.9.1.1
sympy==1.12
thinc==8.1.10
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wasabi==1.1.2
xlrd==2.0.1
xlwt==1.3.0
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.6.0/zh_core_web_sm-3.6.0-py3-none-any.whl#sha256=252fea460a4e63b0d5d81ee2e2339dc69fb21b6b17e2e5c35df75f5cb6ca33d7
